+++ app/src/main/java/com/rushikesh/tasks/data/model/Task.kt
@@ -1,7 +1,7 @@
 package com.rushikesh.tasks.data.model
 
 data class Task(
-    val id: Int? = null,
+    val id: Long? = null,
     val title: String = "",
     val description: String = "",
     val isCompleted: Boolean = false,
